MAINT: better sklearn integration
